#1231: Prevent automatic IDEasy uninstall Fixes: #1231
**Implements**
* Added extra check for IDEasy commandlet to prevent deletion of "corrupted" installation

+++ cli/src/main/java/com/devonfw/tools/ide/tool/LocalToolCommandlet.java
@@ -204,8 +204,11 @@ public abstract class LocalToolCommandlet extends ToolCommandlet {
           return createToolInstallation(installationPath, resolvedVersion, toolVersionFile, false, processContext, extraInstallation);
         }
       } else {
-        this.context.warning("Deleting corrupted installation at {}", installationPath);
-        fileAccess.delete(installationPath);
+        // Makes sure that IDEasy will not delete itself
+        if (!this.tool.equals(IdeasyCommandlet.TOOL_NAME)) {
+          this.context.warning("Deleting corrupted installation at {}", installationPath);
+          fileAccess.delete(installationPath);
+        }
       }
     }
     Path downloadedToolFile = downloadTool(edition, toolRepository, resolvedVersion);
